sphere

+++ model_sdpppg.py
@@ -10,6 +10,7 @@ import pandas as pd
 import os
 import pickle
 from math import log
+from io import BytesIO
 
 from cov import PerObsTemperedOnlineCovariance
 from samplers import ParallelTemperingStickBreakingSampler, bincount2D_vectorized,  \
@@ -565,11 +566,12 @@ class Chain(ParallelTemperingStickBreakingSampler):
         return
 
     def write_to_disk(self, path, nBurn, nThin = 1):
-        folder = os.path.split(path)[0]
-        if not os.path.exists(folder):
-            os.mkdir(folder)
-        if os.path.exists(path):
-            os.remove(path)
+        if type(path) is str:
+            folder = os.path.split(path)[0]
+            if not os.path.exists(folder):
+                os.mkdir(folder)
+            if os.path.exists(path):
+                os.remove(path)
 
         zetas = self.samples.zeta[nBurn :: nThin, 0]
         sigmas = self.samples.sigma[nBurn :: nThin, 0]
@@ -606,8 +608,11 @@ class Chain(ParallelTemperingStickBreakingSampler):
             if hasattr(self.data, attr):
                 out[attr] = self.data.__dict__[attr]
 
-        with open(path, 'wb') as file:
-            pickle.dump(out, file)
+        if type(path) is BytesIO:
+            path.write(pickle.dumps(out))
+        else:
+            with open(path, 'wb') as file:
+                pickle.dump(out, file)
         return
 
     def set_projection(self):
@@ -662,21 +667,15 @@ class Result(object):
     def generate_posterior_predictive_gammas(self, n_per_sample = 1, m = 10):
         new_gammas = []
         for s in range(self.nSamp):
-            njs = np.bincount(
-                self.samples.delta[s], 
-                minlength = int(self.samples.delta[s].max() + 1 + m),
-                )
-            ljs = njs + (njs == 0) * self.samples.eta[s] / m
-            new_zetas = gamma(self.samples.alpha[s], scale = 1 / self.samples.beta[s], size = (m, self.nCol))
-            new_sigmas = np.hstack((
-                np.ones((m, 1)),
-                gamma(self.samples.xi[s], scale = 1 / self.samples.tau[s], size = (m, self.nCol - 1)),
-                ))
-            prob = ljs / ljs.sum()
+            chi = self.samples.chi[s]
+            prob = np.zeros(chi.shape[0])
+            prob += np.log(np.hstack((chi[:-1],(1,))))
+            prob += np.hstack(((0,), np.log(1 - chi[:-1]).cumsum()))
+            np.exp(prob, out = prob)
             deltas = generate_indices(prob, n_per_sample)
-            zeta = np.vstack((self.samples.zeta[s], new_zetas))[deltas]
-            sigma = np.vstack((self.samples.sigma[s], new_sigmas))[deltas]
-            new_gammas.append(gamma(shape = zeta, scale = 1 / sigma))
+            zeta = self.samples.zeta[s][deltas]
+            sigma = self.samples.sigma[s][deltas]
+            new_gammas.append(gamma(shape = zeta, scale = 1/sigma))
         return np.vstack(new_gammas)
 
     def generate_posterior_predictive_hypercube(self, n_per_sample = 1, m = 10):
@@ -696,8 +695,11 @@ class Result(object):
         return
 
     def load_data(self, path):
-        with open(path, 'rb') as file:
-            out = pickle.load(file)
+        if type(path) is BytesIO:
+            out = pickle.loads(path.getvalue())
+        else:
+            with open(path, 'rb') as file:
+                out = pickle.load(file)
         
         deltas = out['deltas']
         zetas  = out['zetas']
